fix/DRY by Sonnar
  - Shared utils:
      - apps/web/lib/ui.ts: STATUS_COLORS, colorFor, isSafeHttpUrl.
      - apps/web/lib/export.ts: exportCsv/exportJson (CSV com escape seguro).
  - Shared component:
      - apps/web/components/LegendChip.tsx: chip de legenda reutilizável.
  - SideMindmap:
      - apps/web/components/SideMindmap.tsx: remove STATUS_COLORS, Legend, isSafeHttpUrl e exports locais; passa a importar de ui.ts/export.ts e usar LegendChip. Mantém grupos e conexões.
  - Mindmap radial:
      - apps/web/components/MindmapPreview.tsx: remove duplicações (STATUS_COLORS, colorFor, isSafeHttpUrl, Legend, exportCsv/Json) e passa a importar os utilitários e LegendChip.

  Efeito

  - Reduz linhas duplicadas e centraliza a lógica de cor, URL seguro, export e legenda.
  - Facilita manutenção e mantém ambos layouts (lateral/radial) consistentes.
